app_server/requirements.txt
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-Nav==0.5
requests==2.12.4
Flask-WTF==0.13.1
Flask-Boto3==0.2.2
Flask-Bcrypt==0.7.1